requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
supabase
python-jose
passlib
bcrypt==4.0.1
python-multipart
fpdf2
openpyxl
pydantic
